feat: addition of skip paths in every workflow templates

+++ test/templates/cluster_endtoend_test_self_hosted.tpl
@@ -13,27 +13,41 @@ jobs:
       - name: Check out code
         uses: actions/checkout@v2
 
+      - name: Check for changes in relevant files
+        uses: dorny/paths-filter@v2
+        id: changes
+        with:
+          filters: |
+            end_to_end:
+              - 'go/**'
+
       - name: Build Docker Image
+        if: steps.changes.outputs.end_to_end == 'true'
         run: docker build -f {{.Dockerfile}} -t {{.ImageName}}:$GITHUB_SHA  .
 
       - name: Run test
+        if: steps.changes.outputs.end_to_end == 'true'
         timeout-minutes: 30
         run: docker run --name "{{.ImageName}}_$GITHUB_SHA" {{.ImageName}}:$GITHUB_SHA /bin/bash -c 'source build.env && go run test.go -keep-data=true -docker=false -print-log -follow -shard {{.Shard}} -- -- --keep-data=true'
 
       - name: Print Volume Used
+        if: steps.changes.outputs.end_to_end == 'true'
         if: ${{"{{ always() }}"}}
         run: |
           docker inspect -f '{{"{{ (index .Mounts 0).Name }}"}}' {{.ImageName}}_$GITHUB_SHA
 
       - name: Cleanup Docker Volume
+        if: steps.changes.outputs.end_to_end == 'true'
         run: |
           docker rm -v {{.ImageName}}_$GITHUB_SHA
 
       - name: Cleanup Docker Container
+        if: steps.changes.outputs.end_to_end == 'true'
         if: ${{"{{ always() }}"}}
         run: |
           docker rm -f {{.ImageName}}_$GITHUB_SHA
 
       - name: Cleanup Docker Image
+        if: steps.changes.outputs.end_to_end == 'true'
         run: |
           docker image rm {{.ImageName}}:$GITHUB_SHA
